Refactor Footer organism
